Adding A350 printer configuration
Largely based on the configuration from mrworf.

Modified the start GCODE to be closer to the start gcode from the forums (https://support.snapmaker.com/hc/en-us/articles/360044341034-What-is-the-recommended-3D-printing-settings-in-Cura-or-Simplify3D-for-Snapmaker-2-0-). I found that the Z axis raise was causing the print to start with the prime still attached.

Modified the end GCODE to just be a simple home, since we don't need to raise so slowly (just pop up on Z and then XY home).