[Filament][Print Settings] Add generation attribution
Noticed that there's a generation attribution at the top of the
combined config files. Probably useful to have this on files generated
to help debug issues with e.g. using SuperSlicer configs on PrusaSlicer,
or if some config manages to stick around long enough without receiving
an update, and would need some compatibility changes for newer versions
of PrusaSlicer/SuperSlicer.