[Print Settings] Proper clearance for extruder head
As discussed [on the
forum](https://forum.snapmaker.com/t/a350-80mm-s-high-speed-acceleration-tests/19155/22?u=nivekmai)
we probably want to be using the 75mm clearance for the extruder when
printing individual items, simply for safety's sake.

Users can always move objects closer if not using the individual items
setting, so this should be fine.